backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
httpx>=0.26.0
sse-starlette>=1.8.2

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9  # For PostgreSQL support

# AI/ML Tools
crewai>=0.13.0
anthropic>=0.8.1
langchain>=0.0.352
langchain-core>=0.1.7
langchain-community>=0.0.9
pydantic>=2.0.0,<3.0.0

# Development Tools
python-dotenv>=1.0.0
setuptools>=69.0.3

# Database and Storage
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
aiofiles>=24.1.0
uvloop>=0.19.0
